move file structure and rename localization.ts to index.ts
